requirements (1).txt
streamlit
requests
beautifulsoup4
readability-lxml
python-docx